Version 0.0.37
0.0.36:

Two new themes have been implemented to improve the look of the graphical user interface: Clarity and Lapis Lazuli.

Encounters can now be sorted by the number of associated effects.

0.0.37:

A new application icon and logo has been created: a ball of yarn with a crochet hook. The icon is visible in the corner of the application window, in the task bar, and in the file manager, as well as in the About window within the SweepWeave application.

The documentation for SweepWeave has been completely revamped. It is now a hypertext document with some interactive elements. The idea is for the documentation to work as an explorable explanation.

The Import from Storyworld feature has been disabled, for now, due to a major bug that I lack sufficient time to fix. It needs to be reworked to check properly for dependencies, to load spools and properties, and to check for collisions.

Some code, and a new scene, have been created to eventually create a more fully functioning find-and-replace feature. I have not had enough time to make much progress on it yet, but I figured out how to highlight text via the _draw function.

Almost ready to release to the wider public.