Version 0.1.4
Version 0.1.4: Ready for release.

Bugfixes:
Loading an encounter from a tab other than the Encounters tab should now select the loaded encounter in the Encounters list on the Encounters tab.
The Graph View should now use the correct edit icon for the clarity theme when first opening SweepWeave.
The switch to exclude self-perceptions from character relationships, which was included in version 0.1.3, introduced a number of errors that showed up when deleting characters or properties. These errors appear to have been fixed successfully in the present version. Relatedly, the process of deleting bnumber pointers from scripts has been simplified so that the removed pointers are replaced with constants rather than different pointers. The script editor has also been changed to hide the bnumber pointer option when no bounded number properties are available to point to.

Optimizations:
The arguments of some function definitions have been edited to include variable types, when possible.

New or Changed Features:
The view menu now includes a toggle to show or hide encounter main text excerpts on the Graph View tab. Encounters have a new function for generating excerpts that stop at the end of a word instead of, potentially, in the middle.
The Graph View has also been updated so that all scripts associated with an encounter will be checked for references to other encounters, where before only boolean outputting scripts were being checked. The graphview should, therefore, now show all script links between encounters. Graph View nodes have also been changed to use the labels "Script Links" and "Next-Page Links" instead of "Prerequisites" and "Consequences."
A number of changes have been made to the spool editing screen. One can now add or remove multiple encounters to or from a spool at once. The addable encounters list will only display encounters that are not already on the current spool. Double-clicking an encounter from the list of encounters on the current spool will switch to the Encounters tab and load that encounter for editing. Finally, the encounters on the current spool will now be sorted by highest to lowest desirability at start of play, and their order can no longer be changed by the author via drag and drop.